Call haplotype specific copy number in single cell datasets

Call haplotype specific copy number in single cell datasets

callHaplotypeSpecificCN(
  CNbins,
  haplotypes,
  eps = 1e-12,
  maskedbins = NULL,
  loherror = 0.02,
  maxCN = NULL,
  selftransitionprob = 0.95,
  progressbar = TRUE,
  ncores = 1,
  phasebyarm = FALSE,
  minfrachaplotypes = 0.7,
  likelihood = "auto",
  minbins = 100,
  minbinschr = 10,
  phased_haplotypes = NULL,
  clustering_method = "copy",
  maxloherror = 0.035,
  mincells = 7,
  overwritemincells = NULL,
  cluster_per_chr = TRUE,
  viterbiver = "cpp",
  filterhaplotypes = 0.1,
  firstpassfiltering = TRUE,
  smoothsingletons = TRUE,
  fillmissing = TRUE,
  global_phasing_for_balanced = FALSE,
  chrs_for_global_phasing = NULL
)

Arguments

CNbins

single cell copy number dataframe with the following columns: cell_id, chr, start, end, state, copy

haplotypes

single cell haplotypes dataframe with the following columns: cell_id, chr, start, end, hap_label, allele1, allele0, totalcounts

eps

default 1e-12

maskedbins

data.frame with columns chr, start and end. These bins will be masked from the inference and copy number states assigned to these bins based on the states of neighbouring bins.

loherror

LOH error rate for initial assignment, this is inferred directly from the data in the second pass, default = 0.02

maxCN

maximum copy number to infer allele specific states, default=NULL which will use the maximum state from CNbins

selftransitionprob

probability to stay in the same state in the HMM, default = 0.95, set to 0.0 for an IID model

progressbar

Boolean to display progressbar or not, default = TRUE, will only show if ncores == 1

ncores

Number of cores to use, default = 1

phasebyarm

Phasing by chromosome arm, default = FALSE

minfrachaplotypes

Minimum proportion of haplotypes to retain when clustering + phasing, default = 0.7

likelihood

Likelihood model for HMM, default is binomial, other option is betabinomial or use auto and the algorithm will choose the likelihood that best fits the data. Default auto

minbins

Minimum number of bins containing both haplotype counts and copy number data for a cell to be included

minbinschr

Minimum number of bins containing both haplotype counts and copy number data per chromosome for a cell to be included

phased_haplotypes

Use this if you want to manually define the haplotypes phasing if for example the default heuristics used by signals does not return a good fit.

clustering_method

Method to use to cluster cells for haplotype phasing, default is copy (using copy column), other option is breakpoints (using breakpoint for clustering)

maxloherror

Maximum value for LOH error rate

mincells

Minimum cluster size used for phasing, default = 7

overwritemincells

Force the number of cells to use for clustering/phasing rather than use the output of the clustering

cluster_per_chr

Whether to cluster per chromosome to rephase alleles or not

filterhaplotypes

filter out haplotypes present in less than X fraction, default is 0.1

firstpassfiltering

Filter out cells with large discrepancy after first pass state assignment

smoothsingletons

Remove singleton bins by smoothing over based on states in adjacent bins

fillmissing

For bins with missing counts fill in values based on neighbouring bins

chrs_for_global_phasing

Which chromosomes to phase using all cells for diploid regions, default is NULL which uses all chromosomes

viterbver

Version of viterbi algorithm to use (cpp or R)

global_phasing_for_diploid

When using cluster_per_chr, use all cells for phasing diploid regions within the cluster

Value

Haplotype specific copy number object

Details

The haplotype specific copy number object include the following additional columns

A A allele copy number

B B allele copy number

state_AS_phased A|B

state_min Minor allele copy number

LOH =LOH if bin is LOH, NO otherwise

state_phase Discretized haplotype specific states

phase Whether the A allele or B allele is dominant

alleleA Counts for the A allele

alleleB Counts for the B allele

totalcounts Total number of counts

BAF B-allele frequency (alleleB / totalcounts)

Examples

sim_data <- simulate_data_cohort(
  clone_num = c(20, 20),
  clonal_events = list(
    list("1" = c(2, 0), "5" = c(3, 1)),
    list("2" = c(6, 3), "3" = c(1, 0))
  ),
  loherror = 0.02,
  coverage = 100
)
#> Joining with `by = join_by(chr, start, end, hap_label)`

results <- callHaplotypeSpecificCN(sim_data$CNbins, sim_data$haplotypes)
#> Filtering out haplotypes present < 10% of cells...
#> Fraction of haplotypes retained after filtering = 1
#> Finding overlapping cell IDs between CN data and haplotype data...
#> Total number of cells in both CN and haplotypes: 40
#> Number of cells in CN data: 40
#> Number of cells in haplotype data: 40
#> Joining bins and haplotypes...
#> Phase haplotypes...
#> Phasing based on distribution across all cells
#> Join phased haplotypes...
#> Reorder haplotypes based on phase...
#> Total number of cells after removing cells with < 100 bins: 40
#> Warning: `progress_estimated()` was deprecated in dplyr 1.0.0.
#> ℹ The deprecated feature was likely used in the signals package.
#>   Please report the issue to the authors.
#> Removing 0 cells for phasing
#> Fitting beta-binomial model to state: 1|1...
#> VGLM    linear loop  1 :  loglikelihood = -377137.2527
#> VGLM    linear loop  2 :  loglikelihood = -352153.843
#> VGLM    linear loop  3 :  loglikelihood = -340882.9118
#> VGLM    linear loop  4 :  loglikelihood = -337650.4394
#> VGLM    linear loop  5 :  loglikelihood = -337002.1207
#> VGLM    linear loop  6 :  loglikelihood = -336893.8146
#> VGLM    linear loop  7 :  loglikelihood = -336876.259
#> VGLM    linear loop  8 :  loglikelihood = -336873.4877
#> Warning: 70103 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  9 :  loglikelihood = -336873.4712
#> Warning: 74773 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  10 :  loglikelihood = -336873.4628
#> Warning: 42506 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  11 :  loglikelihood = -336873.4602
#> Warning: 90739 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  12 :  loglikelihood = -336873.4437
#> Warning: 37465 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  13 :  loglikelihood = -336873.4432
#> Warning: 26134 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  14 :  loglikelihood = -336873.4428
#> Warning: 23344 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  15 :  loglikelihood = -336873.4423
#> Warning: 45896 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  16 :  loglikelihood = -336873.4411
#> Warning: 45445 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  17 :  loglikelihood = -336873.4408
#> Warning: 50944 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  18 :  loglikelihood = -336873.4405
#> Warning: 48113 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  19 :  loglikelihood = -336873.4398
#> Warning: 59871 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  20 :  loglikelihood = -336873.4393
#> Warning: 74076 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  21 :  loglikelihood = -336873.4379
#> Warning: 59876 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  22 :  loglikelihood = -336873.4375
#> Warning: 64436 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  23 :  loglikelihood = -336873.4368
#> Warning: 50430 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  24 :  loglikelihood = -336873.4364
#> Warning: 16700 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  25 :  loglikelihood = -336873.4362
#> Warning: 30431 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  26 :  loglikelihood = -336873.4361
#> Warning: 24150 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  27 :  loglikelihood = -336873.4359
#> Warning: 59627 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  28 :  loglikelihood = -336873.4354
#> Warning: 77082 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  29 :  loglikelihood = -336873.4345
#> Warning: 26699 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
#> VGLM    linear loop  30 :  loglikelihood = -336873.4343
#> Warning: convergence not obtained in 30 IRLS iterations
#> Inferred mean: 0.496, Expected mean: 0.5, Inferred overdispersion (rho): 0
#> Tarones Z-score: -2.202, using binomial model for inference.
#> Using 7 cells for clustering...
#> Clustering chromosome 1
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 2 clusters
#> Distribution of clusters:
#>   Cluster A: 22
#>   Cluster B: 18
#> Clustering chromosome 10
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 11
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 12
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 13
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 14
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 15
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 16
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 17
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 18
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 19
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 2
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 20
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 21
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 22
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 3
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 2 clusters
#> Distribution of clusters:
#>   Cluster A: 23
#>   Cluster B: 17
#> Clustering chromosome 4
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 5
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 2 clusters
#> Distribution of clusters:
#>   Cluster A: 24
#>   Cluster B: 16
#> Clustering chromosome 6
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 7
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 8
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome 9
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome X
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Clustering chromosome Y
#> Creating CN matrix...
#> Calculating UMAP dimensionality reduction...
#> Clustering cells using hdbscan...
#> Identified 1 clusters
#> Distribution of clusters:
#>   Cluster 0: 40
#> Finding overlapping cell IDs between CN data and haplotype data...
#> Total number of cells in both CN and haplotypes: 40
#> Number of cells in CN data: 40
#> Number of cells in haplotype data: 40
#> Joining bins and haplotypes...
#> Phase haplotypes...
#> Join phased haplotypes...
#> Reorder haplotypes based on phase...
#> Total number of cells after removing cells with < 100 bins: 40
#> Average distance from median to expected BAF = 0.0055
#> Average distance from median to expected BAF = 0.0055